--- a/sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OffHeapColumnVector.java
+++ b/sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OffHeapColumnVector.java
@@ -554,6 +554,21 @@ public final class OffHeapColumnVector extends WritableColumnVector {
     return result;
   }
 
+  @Override
+  public int putByteArrays(int rowId, int count, byte[] value, int offset, int length) {
+    reserve(elementsAppended + length * count);
+    int result = 0;
+    for (int i = 0; i < count; i++) {
+      int ret = arrayData().appendBytesWithoutReserveCheck(length, value, offset);
+      int currentRowId = rowId + i;
+      Platform.putInt(null, lengthData + 4L * currentRowId, length);
+      Platform.putInt(null, offsetData + 4L * currentRowId, ret);
+      result += ret;
+    }
+    return result;
+  }
+
+
   // Split out the slow path.
   @Override
   protected void reserveInternal(int newCapacity) {