--- a/sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OnHeapColumnVector.java
+++ b/sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OnHeapColumnVector.java
@@ -532,7 +532,7 @@ public final class OnHeapColumnVector extends WritableColumnVector {
   @Override
   public int putByteArrays(int rowId, int count, byte[] value, int offset, int length) {
     WritableColumnVector child = arrayData();
-    child.reserve(elementsAppended + length * count);
+    child.reserve(elementsAppended + length * count, 1L);
     int result = 0;
     for (int i = 0; i < count; i++) {
       int ret = child.appendBytesWithoutReserveCheck(length, value, offset);