--- a/sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OnHeapColumnVector.java
+++ b/sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OnHeapColumnVector.java
@@ -530,11 +530,11 @@ public final class OnHeapColumnVector extends WritableColumnVector {
   }
 
   @Override
-  public int putByteArrays(int rowId, int count, byte[] value, int offset, int length) {
+  public int putByteArrays(int rowId, int total, byte[] value, int offset, int length) {
     WritableColumnVector child = arrayData();
-    child.reserve(elementsAppended + length * count, 1L);
+    child.reserve(elementsAppended + length * total, 1L);
     int result = 0;
-    for (int i = 0; i < count; i++) {
+    for (int i = 0; i < total; i++) {
       int ret = child.appendBytesWithoutReserveCheck(length, value, offset);
       int currentRowId = rowId + i;
       arrayOffsets[currentRowId] = ret;