@LuciferYang 2022-05-18
Maybe it's better to use a dictionary to store `StringType` partition column. I'm testing it